requirements.txt
colorama==0.4.6
mysql-connector-python==8.3.0
python-dotenv==1.0.1